backend/requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.0
langgraph==0.2.75
faiss-cpu==1.13.2
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.10.6
openai==1.59.8
tiktoken==0.8.0
selenium==4.16.0
webdriver-manager==4.0.1
lxml==5.1.0
PyPDF2==3.0.1
